syncer, sqlmodel(dm): address where handle comments
