+++ dm/syncer/data_validator.go
@@ -833,7 +833,13 @@ func (v *DataValidator) genValidateTableInfo(sourceTable *filter.Table, columnCo
 		// todo: might be connection error, then return error, or downstream table not exists, then set state to stopped.
 		return res, err
 	}
-	pk := downstreamTableInfo.DefaultWhereHandle().UniqueNotNullIdx
+	// if current TI has more columns, clone and strip columns
+	if len(tableInfo.Columns) > columnCount {
+		tableInfo = tableInfo.Clone()
+		tableInfo.Columns = tableInfo.Columns[:columnCount]
+	}
+
+	pk := downstreamTableInfo.WhereHandle(sourceTable, tableInfo).UniqueNotNullIdx
 	if pk == nil {
 		res.message = tableWithoutPrimaryKeyMsg
 		return res, nil
@@ -844,11 +850,6 @@ func (v *DataValidator) genValidateTableInfo(sourceTable *filter.Table, columnCo
 			res.message = downstreamPKColumnOutOfBoundsMsg
 			return res, nil
 		}
-	}
-	// if current TI has more columns, clone and strip columns
-	if len(tableInfo.Columns) > columnCount {
-		tableInfo = tableInfo.Clone()
-		tableInfo.Columns = tableInfo.Columns[:columnCount]
 	}
 
 	res.srcTableInfo = tableInfo
@@ -945,7 +946,7 @@ func (v *DataValidator) processRowsEvent(header *replication.EventHeader, ev *re
 			tableInfo, downstreamTableInfo.TableInfo,
 			nil,
 		)
-		rowChange.SetWhereHandle(downstreamTableInfo.DefaultWhereHandle())
+		rowChange.SetWhereHandle(downstreamTableInfo.WhereHandle(sourceTable, tableInfo))
 		rowChange.SetForeignKeyRelations(downstreamTableInfo.ForeignKeyRelations)
 		size := estimatedRowSize
 		if changeType == rowUpdated && rowChange.IsIdentityUpdated() {
