+++ dm/syncer/syncer.go
@@ -3903,10 +3903,6 @@ func (s *Syncer) precheckForeignKeyReferencedTables(
 	return nil
 }
 
-func basicDownStreamTableInfo(dti *schema.DownstreamTableInfo) *schema.DownstreamTableInfo {
-	return dti.WithoutForeignKeyRelations()
-}
-
 func (s *Syncer) prepareDownStreamTableInfo(
 	tctx *tcontext.Context,
 	sourceTable *filter.Table,
@@ -3923,7 +3919,7 @@ func (s *Syncer) prepareDownStreamTableInfo(
 	}
 
 	if !s.needForeignKeyCausality() {
-		return basicDownStreamTableInfo(dti), nil
+		return dti.WithoutForeignKeyRelations(), nil
 	}
 	if err := s.precheckForeignKeyRouteTopology(tctx.Ctx); err != nil {
 		return nil, err
